Flush digital inputs on startup
This is to avoid spurious increments when the software first starts. It
also happens when there is no connection, because the pull-up resistor on
the input pin registers a digital one. By flushing the input, we ensure that
no interrupt fires until there is actually a rising edge on the pin.