Inversion setting for alarms, initilise alarms correctly.
A separate setting was added to invert the alarm sense. The translation
inversion swaps the meaning of a logical high/low, eg whether the door
is open or close. The alarm inversion setting swaps whether a logical high/low
is a good or a bad thing, eg whether a closed door is bad.

Also properly initialise the alarm path when settings change or when first
configured. Otherwise the alarm is not raised when an alarm condition is present
during startup.