guard against lost interrupts
GPIO reads are edge-triggered on the Venus-GX/Beaglebone and on the Raspberry
Pi. In the event that an interrupt is missed, we still want to pick up on the
level change. We therefore read all the gpios on each iteration, using the poll
object to efficiently react immediately after a level change.