default settings when input type changed
When the input type is changed, reset the other settings to their
defaults. This makes the counter start at zero again and disables
any alarm functionality that may have been attached to the input.

The rate (for pulse counter) and custom name is deliberately left
alone. The rate is only used for the pulse counter and therefore
carries no functionality across to the other functions. It takes
some effort to configure a common name, and a user may well be upset
that he has to start all over.

https://github.com/victronenergy/venus/issues/744