avoid resetting input settings when attrs change
When the maximum value of a setting in localsettings changes,
such as when adding a type as we did in d93397, then the
handler for changing the InputType is called, which on our case
also reset the counter and some of the input settings. Normally
this is correct, changing the input type should start with fresh
settings. In this case it lost the correct settings.

https://github.com/victronenergy/venus/issues/975