Delay service registration until everything is initialised / discovered
This commit delays the service registration until it has performed all the
initialisation and discovery and is about to start the mainloop.

When the service is registered, the GUI and dbus-flashmq will almost
immediately call GetItems. When the registration was done early, because
dbus-digitalinputs was still busy with discovery and startup, it would not
respond, leading to a failed GetItems. This caused the Digital I/O menu to
not show up in the local and the WASM GUI.